Fixed mantis issue #68: Marvell PHY LED setup not quite working as intended
The RX and DUPLEX LEDs have been scrapped. The TX is now AND'ed with itself to go to the jack LED, effectively as a buffer. 
RX and DUPLEX now have testpoints.


Unrelated fix:

JTAG TCK and TRST are pulled up to 3V3AUX as the MAX16071 ICs are powered off of that supply.

Minor tweak:
Altium schematics should now be listed as the same order as page numbers in project view

git-svn-id: https://katfs.kat.ac.za/svnROACH2/elec/roach2@201 f95002ce-95f7-4f7f-9389-3f9229bab526